# docs/swagger.yaml
basePath: /v1/user-services
definitions:
  apperror.Code:
    enum:
    - BAD_REQUEST
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - METHOD_NOT_ALLOWED
    - NOT_ACCEPTABLE
    - REQUEST_TIMEOUT
    - CONFLICT
    - GONE
    - PRECONDITION_FAILED
    - PAYLOAD_TOO_LARGE
    - UNSUPPORTED_MEDIA_TYPE
    - VALIDATION_ERROR
    - UNPROCESSABLE_ENTITY
    - LOCKED
    - FAILED_DEPENDENCY
    - PRECONDITION_REQUIRED
    - TOO_MANY_REQUESTS
    - INTERNAL_ERROR
    - NOT_IMPLEMENTED
    - BAD_GATEWAY
    - SERVICE_UNAVAILABLE
    - TIMEOUT
    - CACHE_MISS
    - CACHE_ERROR
    - MESSAGE_BUS_ERROR
    type: string
    x-enum-varnames:
    - CodeBadRequest
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeMethodNotAllowed
    - CodeNotAcceptable
    - CodeRequestTimeout
    - CodeConflict
    - CodeGone
    - CodePrecondFailed
    - CodePayloadTooLarge
    - CodeUnsupportedMedia
    - CodeValidation
    - CodeUnprocessable
    - CodeLocked
    - CodeFailedDependency
    - CodePrecondRequired
    - CodeTooManyRequest
    - CodeInternal
    - CodeNotImplemented
    - CodeBadGateway
    - CodeServiceUnavail
    - CodeTimeout
    - CodeCacheMiss
    - CodeCacheError
    - CodeMessageBus
  github_com_modami_user-service_internal_dto.AddressResponse:
    properties:
      address_line_1:
        type: string
      address_line_2:
        type: string
      country:
        type: string
      created_at:
        type: string
      district:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      label:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      province:
        type: string
      recipient_name:
        type: string
      ward:
        type: string
    type: object
  github_com_modami_user-service_internal_dto.CreateAddressRequest:
    properties:
      address_line_1:
        maxLength: 512
        type: string
      address_line_2:
        maxLength: 512
        type: string
      country:
        maxLength: 64
        type: string
      district:
        maxLength: 128
        type: string
      is_default:
        type: boolean
      label:
        maxLength: 50
        type: string
      phone:
        type: string
      postal_code:
        maxLength: 20
        type: string
      province:
        maxLength: 128
        type: string
      recipient_name:
        maxLength: 255
        type: string
      ward:
        maxLength: 128
        type: string
    required:
    - address_line_1
    - label
    - phone
    - recipient_name
    type: object
...
  github_com_modami_user-service_internal_dto.RejectKYCRequest:
    properties:
      reason:
        maxLength: 500
        minLength: 1
        type: string
    required:
    - reason
    type: object
...
  github_com_modami_user-service_internal_dto.UpdateAddressRequest:
    properties:
      address_line_1:
        maxLength: 512
        type: string
      address_line_2:
        maxLength: 512
        type: string
      country:
        maxLength: 64
        type: string
      district:
        maxLength: 128
        type: string
      is_default:
        type: boolean
      label:
        maxLength: 50
        type: string
      phone:
        type: string
      postal_code:
        maxLength: 20
        type: string
      province:
        maxLength: 128
        type: string
      recipient_name:
        maxLength: 255
        type: string
      ward:
        maxLength: 128
        type: string
    type: object
...
  github_com_modami_user-service_internal_dto.UpdateStatusRequest:
    properties:
      reason:
        maxLength: 500
        type: string
      status:
        enum:
        - active
        - inactive
        - suspended
        - banned
        type: string
    required:
    - status
    type: object
...
  response.ErrorBody:
    properties:
      code:
        $ref: '#/definitions/apperror.Code'
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      message:
        type: string
    type: object
  response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.Meta:
    properties:
      pagination:
        $ref: '#/definitions/response.Pagination'
      request_id:
        type: string
      timestamp:
        type: integer
    type: object
  response.Pagination:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      page:
        type: integer
      per_page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8086
info:
  contact: {}
  description: User service for the Modami marketplace platform.
  title: Modami User Service API
  version: "1.0"
paths:
  /admin/users:
    get:
      description: List all users with optional search (admin only)
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: 20
        description: Results per page (max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Admin
  /admin/users/{id}/kyc/approve:
    put:
      description: Approve a user's KYC verification (admin only)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Approve KYC
      tags:
      - Admin
  /admin/users/{id}/kyc/reject:
    put:
      consumes:
      - application/json
      description: Reject a user's KYC verification with a reason (admin only)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.RejectKYCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Reject KYC
      tags:
      - Admin
  /admin/users/{id}/status:
    put:
      consumes:
      - application/json
      description: Update a user's account status (admin only)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update user status
      tags:
      - Admin
...
  /users/me/addresses:
    get:
      description: Returns all addresses for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List addresses
      tags:
      - Addresses
    post:
      consumes:
      - application/json
      description: Add a new address for the authenticated user
      parameters:
      - description: Address details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.CreateAddressRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.AddressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Address limit reached
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Add address
      tags:
      - Addresses
  /users/me/addresses/{addr_id}:
    delete:
      description: Delete an address by ID
      parameters:
      - description: Address ID (UUID)
        in: path
        name: addr_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete address
      tags:
      - Addresses
    put:
      consumes:
      - application/json
      description: Update an existing address
      parameters:
      - description: Address ID (UUID)
        in: path
        name: addr_id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.UpdateAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.AddressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update address
      tags:
      - Addresses
  /users/me/addresses/{addr_id}/default:
    put:
      description: Set an address as the default
      parameters:
      - description: Address ID (UUID)
        in: path
        name: addr_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Set default address
      tags:
      - Addresses
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"